requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.4
djangorestframework==3.16.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
idna==3.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.7
pydantic_core==2.33.2
proto-plus==1.26.1
protobuf==5.29.5
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
psutil==5.9.5
memory-profiler==0.61.0